[bug-166147511] change folder structure
