+++ .gitignore
@@ -1,4 +1,5 @@
 /node_modules
 .nyc_output
+.env
 npm-debug.log
 /coverage/
