+++ package.json
@@ -52,6 +52,7 @@
     "chai-http": "^4.3.0",
     "chai-things": "^0.2.0",
     "coveralls": "^3.0.3",
+    "dotenv": "^8.0.0",
     "mocha": "^6.1.4",
     "nodemon": "^1.18.11",
     "nyc": "^14.1.1"
