+++ server/models/usersDb.js
@@ -1,4 +1,14 @@
-
-const users = [];
+const users = [{
+    id:1,
+    firstname: "Patrick",
+    lastname: "Ngabonziza",
+    gender: "Male",
+    address: "Kigali",
+    email: "ngabonzizap7@gmail.com",
+    password: "$2b$10$LHRfSZE43KV0iX38Kv4qKeCYdNAaGOIFUqqsJo4ntuxKyyW0EBuuu",
+    isAdmin: true, 
+    status: "verified",
+    createOn: "2019-May-20 11:05:32"
+}];
 
 export default users;
